+++ src/main/scala/com/codahale/jerkson/Generator.scala
@@ -1,7 +1,7 @@
 package com.codahale.jerkson
 
 import java.io.{File, OutputStream, Writer, StringWriter}
-import org.codehaus.jackson.{JsonGenerator, JsonEncoding}
+import com.fasterxml.jackson.core.{JsonGenerator, JsonEncoding}
 
 trait Generator extends Factory {
   /**
@@ -33,6 +33,11 @@ trait Generator extends Factory {
   def generate[A](obj: A, output: File) {
     generate(obj, factory.createJsonGenerator(output, JsonEncoding.UTF8))
   }
+
+  /**
+   * Returns true if the given class is serializable.
+   */
+  def canSerialize[A](implicit mf: Manifest[A]) = mapper.canSerialize(mf.erasure)
 
   private def generate[A](obj: A, generator: JsonGenerator) {
     generator.writeObject(obj)