+++ src/main/scala/com/codahale/jerkson/Json.scala
@@ -1,13 +1,15 @@
 package com.codahale.jerkson
 
-import org.codehaus.jackson.map.{MappingJsonFactory, ObjectMapper}
-import org.codehaus.jackson.{JsonGenerator, JsonParser => JacksonParser}
+import com.fasterxml.jackson.databind.{MappingJsonFactory, ObjectMapper}
+import com.fasterxml.jackson.core.{JsonGenerator, JsonParser => JacksonParser}
 
 object Json extends Json
 
 trait Json extends Parser with Generator {
+  protected val classLoader = Thread.currentThread().getContextClassLoader
+
   protected val mapper = new ObjectMapper
-  mapper.registerModule(new ScalaModule)
+  mapper.registerModule(new ScalaModule(classLoader))
 
   protected val factory = new MappingJsonFactory(mapper)
   factory.enable(JsonGenerator.Feature.AUTO_CLOSE_JSON_CONTENT)