+++ src/main/scala/com/codahale/jerkson/Parser.scala
@@ -3,8 +3,9 @@ package com.codahale.jerkson
 import io.Source
 import java.net.URL
 import com.codahale.jerkson.AST.{JValue, JNull}
-import org.codehaus.jackson.{JsonNode, JsonParser, JsonProcessingException}
-import org.codehaus.jackson.node.TreeTraversingParser
+import com.fasterxml.jackson.core.{JsonParser, JsonProcessingException}
+import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.node.TreeTraversingParser
 import java.io.{EOFException, Reader, File, InputStream}
 
 trait Parser extends Factory {
@@ -69,12 +70,14 @@ trait Parser extends Factory {
     new StreamingIterator[A](parser, mf)
   }
 
+  /**
+   * Returns true if the given class is deserializable.
+   */
+  def canDeserialize[A](implicit mf: Manifest[A]) = mapper.canDeserialize(Types.build(mapper.getTypeFactory, mf))
+
   private[jerkson] def parse[A](parser: JsonParser, mf: Manifest[A]): A = {
     try {
-      if (mf.erasure == classOf[Option[_]]) {
-        // thanks for special-casing VALUE_NULL, guys
-        Option(parse(parser, mf.typeArguments.head)).asInstanceOf[A]
-      } else if (mf.erasure == classOf[JValue] || mf.erasure == JNull.getClass) {
+      if (mf.erasure == classOf[JValue] || mf.erasure == JNull.getClass) {
         val value: A = parser.getCodec.readValue(parser, Types.build(mapper.getTypeFactory, mf))
         if (value == null) JNull.asInstanceOf[A] else value
       } else {