+++ src/main/scala/com/codahale/jerkson/ScalaModule.scala
@@ -1,21 +1,17 @@
 package com.codahale.jerkson
 
 import deser.ScalaDeserializers
-import org.codehaus.jackson.map.Module.SetupContext
-import org.codehaus.jackson.Version
-import org.codehaus.jackson.map.Module
+import com.fasterxml.jackson.databind.Module.SetupContext
+import com.fasterxml.jackson.core.Version
+import com.fasterxml.jackson.databind.Module
 import ser.ScalaSerializers
 
-/**
- *
- * @author coda
- */
-class ScalaModule extends Module {
-  def version = new Version(0, 2, 0, "")
+class ScalaModule(classLoader: ClassLoader) extends Module {
+  def version = new Version(0, 6, 0, "SNAPSHOT", "com.codahale", "jerkson")
   def getModuleName = "jerkson"
 
   def setupModule(context: SetupContext) {
-    context.addDeserializers(new ScalaDeserializers)
+    context.addDeserializers(new ScalaDeserializers(classLoader, context))
     context.addSerializers(new ScalaSerializers)
   }
 }