+++ src/main/scala/com/codahale/jerkson/Types.scala
@@ -1,7 +1,7 @@
 package com.codahale.jerkson
 
-import org.codehaus.jackson.`type`.JavaType
-import org.codehaus.jackson.map.`type`.{TypeFactory, ArrayType}
+import com.fasterxml.jackson.databind.JavaType
+import com.fasterxml.jackson.databind.`type`.{TypeFactory, ArrayType}
 import scala.collection.JavaConversions.asScalaConcurrentMap
 import java.util.concurrent.ConcurrentHashMap
 
@@ -13,7 +13,7 @@ private[jerkson] object Types {
 
   private def constructType(factory: TypeFactory, manifest: Manifest[_]): JavaType = {
     if (manifest.erasure.isArray) {
-      ArrayType.construct(factory.constructType(manifest.erasure.getComponentType))
+      ArrayType.construct(factory.constructType(manifest.erasure.getComponentType), null, null)
     } else {
       factory.constructParametricType(
         manifest.erasure,