+++ src/main/scala/com/codahale/jerkson/ser/MapSerializer.scala
@@ -1,16 +1,20 @@
 package com.codahale.jerkson.ser
 
-import org.codehaus.jackson.JsonGenerator
-import org.codehaus.jackson.map.{SerializerProvider, JsonSerializer}
-import org.codehaus.jackson.map.annotate.JsonCachable
+import com.fasterxml.jackson.core.JsonGenerator
+import com.fasterxml.jackson.databind.{SerializerProvider, JsonSerializer}
 
-@JsonCachable
 class MapSerializer extends JsonSerializer[collection.Map[_ ,_]] {
-  def serialize(map: collection.Map[_,_], json: JsonGenerator, provider: SerializerProvider) {
-    json.writeStartObject()
-    for ((key, value) <- map) {
-      provider.defaultSerializeField(key.toString, value, json)
+  
+
+    def serialize(map: collection.Map[_,_], json: JsonGenerator, provider: SerializerProvider) {
+        serializeTyped(map, json, provider)
     }
-    json.writeEndObject()
-  }
+
+    private def serializeTyped[A,B](map: collection.Map[A,B], json: JsonGenerator, provider: SerializerProvider) {
+        json.writeStartObject()
+        for ((key, value) <- map) {
+            provider.defaultSerializeField(key.toString, value, json)
+        }
+        json.writeEndObject()
+    }
 }